requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
albumentations>=1.3.0
flask>=3.0.0
waitress>=2.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
kaggle>=1.6.0
pytest>=7.0.0
